software/requirements.txt
Flask
psycopg2-binary
python-dotenv
itsdangerous
selenium
beautifulsoup4
duckduckgo_search
transformers
torch
accelerate>=0.26.0
